[bt][common] AD::DecodeUri handles URIs that start with '\0'
URIs are length-encoded arrays of UTF-8 bytes (more like a std::string
made of UTF-8 bytes), not c-strings. As such, it is possible for a non-
empty URI to contain the byte '\0', which caused an assertion that
code_point >= 2 to fail. Remove this assertion (which is invalid);
instead handle the null byte like any other byte outside the range of
code points which correspond to encoding schemes. This allows for safe
removal of the special-cased uri.empty() check.

Fixed: 69761
Test: `$ fx test bt-host-common-tests`, added a URI with explicit null
      bytes to GAP_AdvertisingDataTest.Uris.
Change-Id: I3236765929f97f80f012cc11556e76fc41db25f3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/485779
Reviewed-by: Ben Lawson <benlawson@google.com>
Commit-Queue: Lucas Jenkins <lucasjenkins@google.com>