[bt][hci] Create transport library
Split out the transport layer functionality from the existing HCI
library into a new transport library. This library defines a
transport layer for sending and receiving command and data packets over
a transport speaking the HCI protocol. It defines tools for working with
command/response transactions, and for working with packet and error
types.

After this CL, the remainder of the existing hci library deals with
managing the state of a controller directed over an HCI transport.

Fixed: 75357
Test: fx test //src/connectivity/bluetooth/core/bt-host
Change-Id: I49189df5310bbff724504aa3f2b1e0654efc5c48
Co-authored-by: Nick Pollard <nickpollard@google.com>
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/519963
Commit-Queue: Ben Lawson <benlawson@google.com>
Reviewed-by: Marie Janssen 💖 <jamuraa@google.com>