[bt][gatt] Handle service changed notifications
Add `removed` and `modified` parameters to
gatt::RemoteServiceManager::ServiceWatcher.

Handle Service Changed notifications in gatt::RemoteServiceManager by
performing service discovery, calculating the difference between
current and new services in the changed range, and notifying the
ServiceWatcher of the service changes.

Service Changed notifications are queued for processing. If a
notification is received during RemoteServiceManager initialization, the
notification is queued and processed as the last step of initialization.

When a service is modified, destroy the old RemoteService and replace
with a new one in order to clean up ongoing procedures and stale
characteristic state. This will also alert clients that the old service
has been invalidated.

Bug: 71986
Test: fx test bt-host-gatt-tests
Change-Id: I0f49efde29603952cd9e3a431ee4c69afc679c57
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/570082
Commit-Queue: Ben Lawson <benlawson@google.com>
Reviewed-by: Xo Wang <xow@google.com>
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>