[bt][gap] Use request/connection tokens in Peer::LowEnergyData
Remove Peer::LowEnergyData::SetConnectionState and replace it with the
following methods:

Add Peer::LowEnergyData::RegisterConnection method, which registers and
returns a Peer::ConnectionToken. Until the token is destroyed, the Peer
is considered connected.

Add Peer::LowEnergyData::RegisterInitializingConnection method, which
registers and returns a Peer::InitializingConnectionToken. Until the
token is destroyed, the Peer is considered initializing (if there are
no ConnectionTokens) or connected (if there are ConnectionTokens).

Instead of explicitly storing the connection state in Peer, calculate it
on demand based on the number of outstanding connection tokens.

Update LowEnergy{Request, Connection} to use connection tokens.

Update PeerFuzzer test to fuzz the new Peer::LowEnergyData methods.

Bug: 70878
Test: fx test bt-host-gap-tests bt-host-fidl-tests
Change-Id: I35233876eec3a781dcdcb9231fd2a60ba6c40913
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/574561
Commit-Queue: Ben Lawson <benlawson@google.com>
Reviewed-by: Lucas Jenkins <lucasjenkins@google.com>