[bt][gap] Re-pair when peer is missing link key
When we believe we are bonded to a peer and initiate authentication, the
controller will ask for our link key and then report Authenticaion
Complete with "PIN or Key Missing" if the peer does not have a link key
for us.  In this case, we should act as if we don't have a link key, and
try to re-pair, but also confirm a new pairing with the local delegate.

Bug: b/210016847
Test: added unit test; fx test bt-host-gap-tests
Test: forget key on peer side, attempt to reconnect from Fuchsia, it
      should succeed.
Change-Id: I70dba3f84eb1e7672f102c79aa950d607c1e1480
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/625662
Fuchsia-Auto-Submit: Marie Janssen <jamuraa@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>
Reviewed-by: Xo Wang <xow@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>