[bt][gap] Coalesce peer name from many sources
Replace SetName() with RegisterName(). Whenever a peer name is read,
register the name along with its source location. Store the name which
originates from the highest priority source in the peer as the canonical
device name.

In AdvertisingData, replace LocalName with a struct that bundles the
local name with a boolean that is true if the name is complete or false
if it is shortened.

Bug: 90213
Test: fx test bt-host-common-tests bt-host-gap-tests
Change-Id: I4361fa591419c3395893c8ed9bd07dc23734995b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/657444
Reviewed-by: Ben Lawson <benlawson@google.com>
Commit-Queue: Ali Saeed <saeedali@google.com>