[bt][core] Remove peer from denylist if we try to connect
When a peer is trying to connect to us after we disconnected due to
local choice, the denylist avoids overzealous autoreconnections from
succeeding.

Unfortunately sometimes we want to connect back to that peer, but they
will not accept our connection and insist on connecting to us.  When
that happens the peer will reject our outgoing connection and continue
to attempt to connect, and we will be stuck in a loop with no way to
connect.

When we attempt to connect via local choice, remove the peer from the
denylist so they can initiate an incoming connection.

Fixed: 102450
Bug: b/228972015
Test: updated unit tests; fx test bt-host-gap-tests
Change-Id: I23f7ca964986a28e3166193e2476835580d2c30f
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/713522
Fuchsia-Auto-Submit: Marie Janssen <jamuraa@google.com>
Reviewed-by: Ben Lawson <benlawson@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>